Fix undefined shift in command/fault/page queue size masks
The queue size mask was computed as (1u << (log2szm1+1)) - 1, and capacity() as
1u << (log2szm1+1). For log2szm1 == 31 the shift is by 32, which is undefined
behavior for a 32-bit unsigned and evaluated to 0 here, forcing the queue tail
(and capacity) to 0 so the queue always appeared empty and no commands were ever
processed. Use a 64-bit shift (1ull) and return capacity() as uint64_t.

This unblocked IODIR.INVAL_DDT processing in the big-endian test, whose command
queue is configured with log2szm1 == 31; without it the DDT cache was never
invalidated and stale device contexts were used after a mode change.

+++ iommu/Iommu.hpp
@@ -147,8 +147,8 @@ namespace TT_IOMMU
     std::array<uint32_t, 2> words;
     uint64_t value;
 
-    unsigned capacity() const {
-        return 1u << (fields.log2szm1+1);
+    uint64_t capacity() const {
+        return 1ull << (fields.log2szm1+1);
     }
   };
